Used custom hook, avoid code duplications
