+++ backend/data/user-places.json
@@ -1,1 +1,1 @@
-[{"id":"p2","title":"Sahara Desert Dunes","image":{"src":"desert-dunes.jpg","alt":"Golden dunes stretching to the horizon in the Sahara Desert."},"lat":25,"lon":0},{"id":"p10","title":"Parisian Streets","image":{"src":"parisian-streets.jpg","alt":"Charming streets of Paris with historic buildings and cafes."},"lat":48.8566,"lon":2.3522},{"id":"p9","title":"Great Barrier Reef","image":{"src":"great-barrier-reef.jpg","alt":"Vibrant coral formations of the Great Barrier Reef underwater."},"lat":-18.2871,"lon":147.6992},{"id":"p1","title":"Forest Waterfall","image":{"src":"forest-waterfall.jpg","alt":"A tranquil forest with a cascading waterfall amidst greenery."},"lat":44.5588,"lon":-80.344},{"id":"p4","title":"Caribbean Beach","image":{"src":"caribbean-beach.jpg","alt":"Pristine white sand and turquoise waters of a Caribbean beach."},"lat":18.2208,"lon":-66.5901},{"id":"p14","title":"Kerala Backwaters","image":{"src":"kerala-backwaters.jpg","alt":"Tranquil waters and lush greenery of the Kerala backwaters."},"lat":9.4981,"lon":76.3388},{"id":"p13","title":"Taj Mahal","image":{"src":"taj-mahal.jpg","alt":"The iconic Taj Mahal reflecting in its surrounding waters during sunrise."},"lat":27.1751,"lon":78.0421}]
+[{"id":"p8","title":"Japanese Temple","image":{"src":"japanese-temple.jpg","alt":"Ancient Japanese temple surrounded by autumn foliage."},"lat":34.9949,"lon":135.785},{"id":"p4","title":"Caribbean Beach","image":{"src":"caribbean-beach.jpg","alt":"Pristine white sand and turquoise waters of a Caribbean beach."},"lat":18.2208,"lon":-66.5901},{"id":"p3","title":"Himalayan Peaks","image":{"src":"majestic-mountains.jpg","alt":"The sun setting behind snow-capped peaks of majestic mountains."},"lat":27.9881,"lon":86.925},{"id":"p10","title":"Parisian Streets","image":{"src":"parisian-streets.jpg","alt":"Charming streets of Paris with historic buildings and cafes."},"lat":48.8566,"lon":2.3522},{"id":"p9","title":"Great Barrier Reef","image":{"src":"great-barrier-reef.jpg","alt":"Vibrant coral formations of the Great Barrier Reef underwater."},"lat":-18.2871,"lon":147.6992},{"id":"p14","title":"Kerala Backwaters","image":{"src":"kerala-backwaters.jpg","alt":"Tranquil waters and lush greenery of the Kerala backwaters."},"lat":9.4981,"lon":76.3388}]
